docs: Add guide on how to use Git
The project includes a set of .md files with information that covers the
following areas:

1 - Git installation on a Windows environment
2 - Creation of new repositories
3 - Basic commands required to use Git on a daily basis
4 - Good practices to make better use of this tool

This documentation doesn't cover any advanced features such as rebasing
or submodules, for example. The guide is meant to be used by people who
are new to Git, and so it only explains what's strictly necessary to get
started.